% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cog_data}
\alias{cog_data}
\title{Cognition Data}
\format{
A dataframe with 100 rows and 19 variables:
\describe{
\item{HHID}{Household identifier, a unique 6-digit integer.}
\item{PN}{Person number, a unique 1- or 2-digit integer within each household.}

\item{SD182M1}{Immediate word recall test score for the first word.}
\item{SD182M2}{Immediate word recall test score for the second word.}
\item{SD182M3}{Immediate word recall test score for the third word.}
\item{SD182M4}{Immediate word recall test score for the fourth word.}
\item{SD182M5}{Immediate word recall test score for the fifth word.}

\item{SD183M1}{Delayed word recall test score for the first word.}
\item{SD183M2}{Delayed word recall test score for the second word.}
\item{SD183M3}{Delayed word recall test score for the third word.}
\item{SD183M4}{Delayed word recall test score for the fourth word.}
\item{SD183M5}{Delayed word recall test score for the fifth word.}

\item{SD142}{Serial subtraction, result of subtracting 7 from 100.}
\item{SD143}{Serial subtraction, result of subtracting 7 from the previous number.}
\item{SD144}{Serial subtraction, result of subtracting 7 from the previous number.}
\item{SD145}{Serial subtraction, result of subtracting 7 from the previous number.}
\item{SD146}{Serial subtraction, result of subtracting 7 from the previous number.}

\item{SD124}{Backwards counting test, success on the first attempt (1 = success, 0 = fail).}
\item{SD129}{Backwards counting test, success on the second attempt (1 = success, 0 = fail).}
}
}
\usage{
cog_data
}
\description{
A simulated dataset with cognition test scores, following the same methodology as the Health and Retirement Study (HRS).
The dataset includes immediate word recall, delayed word recall, serial subtraction, and backwards counting tasks, with scores representing
cognitive performance on these tests.
}
\examples{
# Load the data
data(cog_data)

# View the first few rows
head(cog_data)

}
\keyword{datasets}
